requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
Django==4.2.27
django-ninja==1.5.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
sqlparse==0.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
